// src/lib/api/generated/schemas/prescriptionsControllerFindAll200.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Prescription Management API
 * API documentation for the MVP Prescription Management System.
 * OpenAPI spec version: 1.0
 */
import type { PaginatedResultDto } from './paginatedResultDto';
import type { PaginationMetaDto } from './paginationMetaDto';
import type { PrescriptionResponseDto } from './prescriptionResponseDto';

export type PrescriptionsControllerFindAll200 = PaginatedResultDto & {
  data?: PrescriptionResponseDto[];
  meta?: PaginationMetaDto;
};